flask_app/requirements.txt
Flask
Flask-CORS
Flask-JWT-Extended
pymongo
python-dotenv
werkzeug
